Fixed use of wait -n since it is not POSIX.
Process management is with a list of PIDs. At the end of the main
process, we iterate the list of PIDs to wait for each one and
check if they exited normally or not. Currently, there is no
implementation for limiting to a fixed number of processes like
before. That feature is planned.