Implemented process managment.
Main process will hang out in a loop when max jobs are running. While in
the loop, it will poll each PID and wait on them as they terminate.
After waiting on a terminated PID, it will remove a PID from the list so
the next URL in the list can download.